crud head of family done

+++ app/Repositories/HeadOfFamilyRepository.php
@@ -6,6 +6,7 @@ use App\Interfaces\HeadOfFamilyRepositoryInterface;
 use App\Models\HeadOfFamily;
 use App\Models\User;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 
 class HeadOfFamilyRepository implements HeadOfFamilyRepositoryInterface
 {
@@ -42,16 +43,16 @@ class HeadOfFamilyRepository implements HeadOfFamilyRepositoryInterface
         DB::beginTransaction();
 
         try {
-            //code...
-            $user = new User();
-            $user->name = $data['name'];
-            $user->email = $data['email'];
-            $user->password = bcrypt($data['password']);
-            $user->save();
+            $userRepository = new UserRespository();
+            $user = $userRepository->create([
+                'name' => $data['name'],
+                'email' => $data['email'],
+                'password' => bcrypt($data['password']),
+            ]);
 
             $headOfFamily = new HeadOfFamily();
             $headOfFamily->user_id = $user->id;
-            $headOfFamily->profile_picture = $data['profile_picture'];
+            $headOfFamily->profile_picture = $data['profile_picture']->store('assets/head-of-family', 'public');
             $headOfFamily->identity_number = $data['identity_number'];
             $headOfFamily->gender = $data['gender'];
             $headOfFamily->date_of_birth = $data['date_of_birth'];
@@ -82,7 +83,13 @@ class HeadOfFamilyRepository implements HeadOfFamilyRepositoryInterface
         try {
             //code...
             $headOfFamily = $this->getById($id);
-            $headOfFamily->profile_picture = $data['profile_picture'];
+            if(isset($data['profile_picture'])){
+                if($headOfFamily->profile_picture){
+                    // delete image
+                    Storage::disk('public')->delete($headOfFamily->profile_picture);
+                }
+                $headOfFamily->profile_picture = $data['profile_picture']->store('assets/head-of-family', 'public');
+            }
             $headOfFamily->identity_number = $data['identity_number'];
             $headOfFamily->gender = $data['gender'];
             $headOfFamily->date_of_birth = $data['date_of_birth'];
@@ -90,6 +97,14 @@ class HeadOfFamilyRepository implements HeadOfFamilyRepositoryInterface
             $headOfFamily->occupation = $data['occupation'];
             $headOfFamily->marital_status = $data['marital_status'];
             $headOfFamily->save();
+
+            $userRepository = new UserRespository();
+            $userRepository->update([
+                'name' => $data['name'],
+                'email' => $data['email'],
+                'password' => isset($data['password']) ? bcrypt($data['password'])
+                    : $headOfFamily->user->password,
+            ], $headOfFamily->user_id);
 
             DB::commit();
             return $headOfFamily;
@@ -106,7 +121,11 @@ class HeadOfFamilyRepository implements HeadOfFamilyRepositoryInterface
         try {
             //code...
             $headOfFamily = $this->getById($id);
+            if(!$headOfFamily){
+                throw new \Exception('Head of family not found');
+            }
             $headOfFamily->delete();
+            $headOfFamily->profile_picture ? Storage::disk('public')->delete($headOfFamily->profile_picture) : null;
             DB::commit();
             return true;
         } catch (\Throwable $th) {
